refactoring + linting
