+++ src/main/kotlin/org/compiere/bo/X_C_Opportunity.kt
@@ -24,7 +24,6 @@ import java.util.Properties
 
 import org.compiere.model.I_C_Opportunity
 import org.compiere.orm.MTable
-import org.compiere.orm.PO
 import org.compiere.orm.BasePO
 import org.idempiere.orm.I_Persistent
 import org.idempiere.common.util.Env
