+++ src/test/kotlin/company/bigger/test/BPartnerTests.kt
@@ -1,0 +1,18 @@
+package company.bigger.test
+
+import company.bigger.test.support.BaseTest
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.junit4.SpringRunner
+
+@RunWith(SpringRunner::class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureMockMvc
+class BPartnerTests : BaseTest() {
+
+    @Test
+    fun contextLoads() {
+    }
+}
