+++ src/test/kotlin/company/bigger/test/IdempiereMicroSpringApplicationTests.kt
@@ -1,84 +1,23 @@
 package company.bigger.test
 
-import company.bigger.web.controller.UserController
-import company.bigger.test.support.BaseTest
-import company.bigger.Micro
-import company.bigger.dto.UserLoginModel
-import company.bigger.service.UserService
-import company.bigger.test.clients.LoginClient
-import company.bigger.test.clients.UserClient
+import company.bigger.test.support.BaseIntegrationTest
 import company.bigger.test.support.randomString
-import feign.Feign
 import feign.FeignException
-import feign.gson.GsonDecoder
-import feign.gson.GsonEncoder
-import org.junit.Before
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.core.env.Environment
-import org.springframework.test.context.junit4.SpringRunner
-import org.springframework.boot.web.server.LocalServerPort
-import kotlin.test.*
+import kotlin.test.assertTrue
+import kotlin.test.assertFalse
+import kotlin.test.assertNotNull
+import kotlin.test.assertEquals
+import kotlin.test.fail
 
-@RunWith(SpringRunner::class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureMockMvc
-class IdempiereMicroSpringApplicationTests : BaseTest() {
+class IdempiereMicroSpringApplicationTests : BaseIntegrationTest() {
 
     @Test
     fun contextLoads() {
     }
 
-    @Autowired
-    private lateinit var userService: UserService
     @Test
-    fun `GardenUser can login (service)`() {
-        val result = userService.login(UserLoginModel("GardenUser", "GardenUser"))
-        assertNotNull(result?.token)
-    }
-
-    @Autowired
-    private lateinit var userController: UserController
-
-    @Autowired
-    private lateinit var micro: Micro
-
-    @Before
-    override fun prepare() {
-        super.prepare()
-        micro.startup()
-        loginClient = buildClient(LoginClient::class.java)
-        userClient = buildClient(UserClient::class.java)
-    }
-
-    @Test
-    fun `GardenUser can login (controller)`() {
-        val result = userController.login("GardenUser", "GardenUser")
-        assertNotNull(result?.token)
-    }
-
-    @Autowired
-    protected lateinit var environment: Environment
-
-    @LocalServerPort
-    var randomServerPort: Int = 0
-
-    private fun <T> buildClient(t: Class<T>): T {
-        environment.let {
-            return Feign.builder()
-                    .encoder(GsonEncoder()).decoder(GsonDecoder())
-                    .target(t, "http://localhost:$randomServerPort")
-        }
-    }
-
-    protected var loginClient: LoginClient? = null
-    protected var userClient: UserClient? = null
-
-    @Test
-    fun `GardenUser can login (REST)`() {
+    fun `GardenUser can login`() {
         val gardenUserLogin = loginClient?.login("GardenUser", "GardenUser")
         println("$gardenUserLogin")
         assertNotNull(gardenUserLogin)
@@ -115,7 +54,7 @@ class IdempiereMicroSpringApplicationTests : BaseTest() {
     }
 
     @Test
-    fun `Random user can not login (REST)`() {
+    fun `Random user can not login`() {
         val gardenUserLogin = loginClient?.login(randomString(10), randomString(10))
         println("$gardenUserLogin")
         assertNotNull(gardenUserLogin)
@@ -128,9 +67,8 @@ class IdempiereMicroSpringApplicationTests : BaseTest() {
         val token = randomString(50)
         try {
             userClient?.profile(token)
-
         } catch (ex: FeignException) {
-            if (ex.status()!=401) fail("Should HTTP 401 fail, not $ex")
+            if (ex.status() != 401) fail("Should HTTP 401 fail, not $ex")
         }
     }
 }
