+++ src/test/kotlin/company/bigger/test/support/BaseTest.kt
@@ -1,6 +1,7 @@
 package company.bigger.test.support
 
 import company.bigger.common.db.CConnection
+import company.bigger.service.UserService
 import company.bigger.util.DatabaseImpl
 import company.bigger.util.DummyEventManager
 import company.bigger.util.DummyService
@@ -9,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.junit4.SpringRunner
 import company.bigger.util.Ini
-import org.compiere.bo.tests.TestUpdateCustomer
 import org.idempiere.common.db.Database
 import org.idempiere.common.util.CLogger
 import org.idempiere.common.util.DB
@@ -27,11 +27,14 @@ abstract class BaseTest {
     @Autowired
     protected lateinit var cconnection: ICConnection
 
+    @Autowired
+    protected lateinit var userService: UserService
+
     @Before
     open fun prepare() {
         DummyService.setup()
         DummyEventManager.setup()
-        CLogger.getCLogger(TestUpdateCustomer::class.java)
+        CLogger.getCLogger(BaseTest::class.java)
         val db = Database()
         db.setDatabase(DatabaseImpl())
         DB.setDBTarget(CConnection.get())
